feat: rename loginController to authController
