+++ routes/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const usersController = require('../controllers/usersController');
-const loginController = require('../controllers/loginController');
+const authController = require('../controllers/authController');
 const familyGroupsController = require('../controllers/familyGroupsController');
 const petsController = require('../controllers/petsController');
 const fedsController = require('../controllers/fedsController');
@@ -9,10 +9,10 @@ const auth = require('../middlewares/authentication');
 const api = express.Router();
 
 // Login
-api.post('/signup', loginController.signUp);
-api.post('/signin', loginController.signIn);
-api.post('/signingoogle', loginController.signInGoogle);
-// api.get('/signingithub', loginController.signInGitHub); // TODO: StandBy
+api.post('/signup', authController.signUp);
+api.post('/signin', authController.signIn);
+api.post('/signingoogle', authController.signInGoogle);
+// api.get('/signingithub', authController.signInGitHub); // TODO: StandBy
 
 // Users
 api.get('/getusers', auth, usersController.getUsers);
